+++ tests/platform_helper/domain/test_conduit.py
@@ -4,11 +4,11 @@ import pytest
 
 from dbt_platform_helper.domain.conduit import Conduit
 from dbt_platform_helper.exceptions import AddonNotFoundError
+from dbt_platform_helper.exceptions import CreateTaskTimeoutError
 from dbt_platform_helper.exceptions import InvalidAddonTypeError
+from dbt_platform_helper.exceptions import NoClusterError
 from dbt_platform_helper.exceptions import ParameterNotFoundError
-from dbt_platform_helper.providers.copilot import CreateTaskTimeoutError
-from dbt_platform_helper.providers.copilot import NoClusterError
-from dbt_platform_helper.providers.secrets import SecretNotFoundError
+from dbt_platform_helper.exceptions import SecretNotFoundError
 from dbt_platform_helper.utils.application import Application
 from dbt_platform_helper.utils.application import Environment
