+++ tests/platform_helper/providers/test_copilot.py
@@ -8,15 +8,15 @@ from botocore.exceptions import ClientError
 from moto import mock_aws
 
 from dbt_platform_helper.exceptions import AddonNotFoundError
+from dbt_platform_helper.exceptions import NoClusterError
 from dbt_platform_helper.exceptions import ParameterNotFoundError
 from dbt_platform_helper.providers.copilot import CreateTaskTimeoutError
-from dbt_platform_helper.providers.copilot import NoClusterError
-from dbt_platform_helper.providers.copilot import addon_client_is_running
 from dbt_platform_helper.providers.copilot import connect_to_addon_client_task
 from dbt_platform_helper.providers.copilot import create_addon_client_task
 from dbt_platform_helper.providers.copilot import create_postgres_admin_task
-from dbt_platform_helper.providers.copilot import get_cluster_arn
-from dbt_platform_helper.providers.copilot import get_or_create_task_name
+from dbt_platform_helper.providers.ecs import addon_client_is_running
+from dbt_platform_helper.providers.ecs import get_cluster_arn
+from dbt_platform_helper.providers.ecs import get_or_create_task_name
 from dbt_platform_helper.providers.secrets import SecretNotFoundError
 from dbt_platform_helper.providers.secrets import (
     _normalise_secret_name as normalise_secret_name,
