+++ tests/platform_helper/providers/test_secrets.py
